+++ src/config.ml
@@ -442,6 +442,7 @@ let valid_settings = [
   "doctype"; "keep_doctype";
   "index_page"; "index_file";
   "clean_urls"; "clean_url_trailing_slash"; "clean_urls_exclude_regexes";
+  "site_url";
   "page_file_extensions"; "ignore_extensions"; "default_extension"; "keep_extensions";
   "markdown_extensions";
   "markdown_strict_commonmark";
@@ -494,6 +495,7 @@ let _update_settings settings config =
        clean_urls = find_bool_or ~default:settings.clean_urls st ["clean_urls"];
        clean_url_trailing_slash = find_bool_or ~default:settings.clean_url_trailing_slash st ["clean_url_trailing_slash"];
        clean_url_exclude_regexes = find_strings_or ~default:[] st ["clean_urls_exclude_regexes"];
+       site_url = OH.find_string_opt st ["site_url"];
        page_extensions = find_strings_or ~default:settings.page_extensions st ["page_file_extensions"];
        markdown_extensions = find_strings_or ~default:settings.markdown_extensions st ["markdown_extensions"];
        markdown_strict_commonmark = find_bool_or ~default:settings.markdown_strict_commonmark st ["markdown_strict_commonmark"];